# fix: memory leak during STEK rotation