--- a/tls/s2n_config.h
+++ b/tls/s2n_config.h
@@ -32,7 +32,6 @@
 #include "utils/s2n_set.h"
 
 #define S2N_MAX_TICKET_KEYS       48
-#define S2N_MAX_TICKET_KEY_HASHES 500 /* 10KB */
 
 /*
  * TLS1.3 does not allow alert messages to be fragmented, and some TLS
@@ -133,8 +132,7 @@ struct s2n_config {
 
     uint64_t session_state_lifetime_in_nanos;
 
-    struct s2n_set *ticket_keys;
-    struct s2n_set *ticket_key_hashes;
+    struct s2n_array *ticket_keys;
     uint64_t encrypt_decrypt_key_lifetime_in_nanos;
     uint64_t decrypt_key_lifetime_in_nanos;